+++ app/src/main/java/com/amora/app/dialog/EditProfileBottomSheetDialog.java
@@ -49,24 +49,20 @@ public class EditProfileBottomSheetDialog extends BottomSheetDialog implements A
 
     @Override
     public void onClick(View view) {
-        switch (view.getId()) {
-            case R.id.check:
-                new PictureView(getContext(), pic, selectedPosition);
-                dismiss();
-                return;
+        int id = view.getId();
 
-            case R.id.delete:
-                apiManager.deleteProfileImage(String.valueOf(pic.get(selectedPosition).getId()));
-                return;
+        if (id == R.id.check) {
+            new PictureView(getContext(), pic, selectedPosition);
+            dismiss();
 
-            case R.id.set_profile_pic:
+        } else if (id == R.id.delete) {
+            apiManager.deleteProfileImage(String.valueOf(pic.get(selectedPosition).getId()));
 
-                apiManager.setProfilePicture(String.valueOf(pic.get(selectedPosition).getId()));
-                return;
+        } else if (id == R.id.set_profile_pic) {
+            apiManager.setProfilePicture(String.valueOf(pic.get(selectedPosition).getId()));
 
-            case R.id.cancel:
-                dismiss();
-                return;
+        } else if (id == R.id.cancel) {
+            dismiss();
         }
     }
